Move all logic into static buildFrom method in Scorecard Model
