+++ api/lib/domain/models/Scorecard.js
@@ -1,3 +1,4 @@
+const _ = require('lodash');
 const constants = require('../constants');
 
 class Scorecard {
@@ -24,6 +25,36 @@ class Scorecard {
     this.level = this._getCompetenceLevel(roundedEarnedPix);
     this.pixScoreAheadOfNextLevel = this._getpixScoreAheadOfNextLevel(roundedEarnedPix);
     this.status = status;
+  }
+
+  static buildFrom({ userId, userKEList, competence, competenceEvaluations }) {
+    const sortedKEGroupedByCompetence = _.groupBy(userKEList, 'competenceId');
+    const knowledgeElementsOfCompetence = sortedKEGroupedByCompetence[competence.id];
+    const totalEarnedPixByCompetence = _.sumBy(knowledgeElementsOfCompetence, 'earnedPix');
+
+    return new Scorecard({
+      id: `${userId}_${competence.id}`,
+      name: competence.name,
+      description: competence.description,
+      competenceId: competence.id,
+      index: competence.index,
+      area: competence.area,
+      earnedPix: totalEarnedPixByCompetence,
+      status: Scorecard._computeStatus(knowledgeElementsOfCompetence, competence.id, competenceEvaluations)
+    });
+  }
+
+  static _computeStatus(knowledgeElements, competenceId, competenceEvaluations) {
+    if (_.isEmpty(knowledgeElements)) {
+      return 'NOT_STARTED';
+    }
+
+    const competenceEvaluationForCompetence = _.find(competenceEvaluations, { competenceId });
+    const stateOfAssessment = _.get(competenceEvaluationForCompetence, 'assessment.state');
+    if (stateOfAssessment === 'completed') {
+      return 'COMPLETED';
+    }
+    return 'STARTED';
   }
 
   _getCompetenceLevel(earnedPix) {
